Buggy (!) version of only second network refinement using split heuristic ideas from Bak.
This is work in progress! In particular we are strugglying with an assertion error in the check/refinement function which claims that even though the counter-example is spurious there are no over-approximations left.
The error only shows up after a long run time on the ACAS Xu 1010 property and might hint to a larger problem with zonostate handling? Are we doing duplicate work? dropping things?